move preview to a separate command
